other pages (#576)
* other pages

* fix mobile bug

* other fix

* fix ui bugs

* fix testnet
