+++ .env.example
@@ -1,3 +1,3 @@
 NEXT_PUBLIC_HOMESERVER=ufibwbmed6jeq9k4p583go95wofakh9fwpp4k734trq79pd9u1uy
 NEXT_PUBLIC_NEXUS=https://nexus.staging.pubky.app
-TESTNET=
+NEXT_PUBLIC_TESTNET=
