+++ apps/web/app/onboarding/components/_Content.tsx
@@ -1,0 +1,184 @@
+'use client';
+
+import Link from 'next/link';
+import { useEffect, useState } from 'react';
+import { Header, Content, Typography, Button, Icon } from '@social/ui-shared';
+import { useRouter } from 'next/navigation';
+import { usePubkyClientContext } from '@/contexts';
+import { Links } from '@/types/Post';
+
+export default function Index() {
+  const [isMobile, setIsMobile] = useState(false);
+  const { pubky, signUp, isLoggedIn } = usePubkyClientContext();
+  const router = useRouter();
+  const [logoLink, setLogoLink] = useState('/onboarding');
+  const [loading, setLoading] = useState(false);
+  const links: Links[] = [];
+
+  useEffect(() => {
+    const checkMobile = () => {
+      setIsMobile(window.innerWidth < 640);
+    };
+
+    checkMobile();
+    window.addEventListener('resize', checkMobile);
+
+    return () => window.removeEventListener('resize', checkMobile);
+  }, []);
+
+  useEffect(() => {
+    async function fetchData() {
+      const loggedIn = await isLoggedIn();
+      if (!loggedIn) {
+        setLogoLink('/onboarding');
+      } else {
+        setLogoLink('/home');
+      }
+    }
+    fetchData();
+  }, [pubky, isLoggedIn]);
+
+  const handleSubmit = async () => {
+    if (loading) {
+      return;
+    }
+    try {
+      setLoading(true);
+
+      const signUpResponse = await signUp({
+        name: '',
+        bio: '',
+        image: '/images/webp/Userpic.webp',
+        links: links,
+      });
+
+      if (!signUpResponse) {
+        throw new Error('Something went wrong');
+      }
+
+      router.push('/home');
+    } catch (error) {
+      console.log(error);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <Content.Main>
+      <Header.Root className="backdrop-blur-[0px]">
+        <Header.Logo link={logoLink} />
+        <div className="flex gap-12 items-center">
+          <div className="h-6 justify-start items-start gap-6 sm:inline-flex hidden">
+            <Link
+              target="_blank"
+              href="https://github.com/pubky"
+              className="cursor-pointer opacity-30 hover:opacity-100"
+            >
+              <Icon.Github size="24" />
+            </Link>
+            <Link
+              target="_blank"
+              href="https://x.com/getpubky"
+              className="cursor-pointer opacity-30 hover:opacity-100"
+            >
+              <Icon.Twitter size="24" />
+            </Link>
+            <Link
+              target="_blank"
+              href="https://www.youtube.com/channel/UCyNruUjynpzvQXNTxbJBLmg"
+              className="cursor-pointer opacity-30 hover:opacity-100"
+            >
+              <Icon.Youtube width="24" height="24" />
+            </Link>
+          </div>
+          <Header.Action
+            icon={<Icon.SignIn size="16" />}
+            id="onboarding-sign-in-btn"
+          >
+            Sign in
+          </Header.Action>
+        </div>
+      </Header.Root>
+      <Content.Grid className="relative z-20 xl:mt-14">
+        <Typography.Display className="text-5xl sm:text-7xl xl:text-9xl xl:leading-[128px]">
+          Unlock <br />
+          the web.
+        </Typography.Display>
+        <Typography.Body
+          variant="large"
+          className="text-[22px] sm:text-2xl leading-tight text-opacity-50 mt-2 sm:mt-4"
+        >
+          Your keys, your content, your rules.
+        </Typography.Body>
+        <div className="relative flex gap-3 mt-6 sm:mt-12">
+          <Link href="/onboarding/intro">
+            <Button.Large
+              className="bg-[#c8ff00] border-[#c8ff00]"
+              colorText="text-[#c8ff00]"
+            >
+              Create Account
+            </Button.Large>
+          </Link>
+          <Button.Large
+            onClick={!loading ? () => handleSubmit() : undefined}
+            variant="secondary"
+            className="w-auto"
+            loading={loading}
+          >
+            Explore Pubky
+          </Button.Large>
+          <Button.Large
+            onClick={() => window.open('https://pubky.org', '_blank')}
+            variant="secondary"
+            className="w-auto hidden sm:flex"
+          >
+            Pubky Core
+          </Button.Large>
+        </div>
+        <div className="fixed bottom-12">
+          <div className="mb-2 h-6 justify-start items-start gap-6 inline-flex sm:hidden">
+            <Link
+              target="_blank"
+              href="https://github.com/pubky"
+              className="cursor-pointer opacity-30 hover:opacity-100"
+            >
+              <Icon.Github size="24" />
+            </Link>
+            <Link
+              target="_blank"
+              href="https://x.com/getpubky"
+              className="cursor-pointer opacity-30 hover:opacity-100"
+            >
+              <Icon.Twitter size="24" />
+            </Link>
+            <Link
+              target="_blank"
+              href="https://www.youtube.com/channel/UCyNruUjynpzvQXNTxbJBLmg"
+              className="cursor-pointer opacity-30 hover:opacity-100"
+            >
+              <Icon.Youtube width="24" height="24" />
+            </Link>
+          </div>
+          <Typography.Body
+            variant="small"
+            className="text-[13.5px] text-opacity-30 font-normal"
+          >
+            Synonym Software Ltd. ©2024.
+          </Typography.Body>
+        </div>
+      </Content.Grid>
+      <div className="w-full">
+        <div
+          style={{
+            backgroundImage: isMobile
+              ? "url('/images/webp/home-mobile.webp')"
+              : "url('/images/webp/home.webp')",
+            marginTop: isMobile ? '150px' : '',
+          }}
+          className="fixed inset-0 bg-cover bg-center pointer-events-none"
+        />
+      </div>
+    </Content.Main>
+  );
+}
