+++ apps/web/app/onboarding/components/index.tsx
@@ -1,5 +1,7 @@
+import Content from './_Content'
 import Layout from './_Layout';
 
 export const Onboarding = {
+  Content,
   Layout,
 };
