+++ apps/web/app/onboarding/confirm/components/index.tsx
@@ -1,0 +1,5 @@
+import Content from './_Content';
+
+export const Confirm = {
+  Content,
+};
