+++ apps/web/app/onboarding/pubky/components/index.tsx
@@ -1,0 +1,5 @@
+import Content from './_Content';
+
+export const Pubky = {
+  Content,
+};
