+++ apps/web/app/onboarding/register/components/index.tsx
@@ -1,0 +1,5 @@
+import Content from './_Content';
+
+export const Register = {
+  Content,
+};
