+++ apps/web/app/onboarding/sign-in/page.tsx
@@ -1,23 +1,11 @@
-import { Typography } from '@social/ui-shared';
-import { Onboarding } from '../components';
-import { Card } from './Card';
+import { getSeoMetadata } from '@/components/HeaderSEO';
+import { SignIn } from './components/index';
+
+export const metadata = getSeoMetadata({
+  title: 'SignIn | Pubky.app',
+  description: 'Pubky.app - Unlock the web.',
+});
 
 export default function Index() {
-  return (
-    <Onboarding.Layout currentStep={1}>
-      <Typography.Display>Let&apos;s get started</Typography.Display>
-      <Typography.Body
-        variant="large"
-        className="text-[22px] sm:text-2xl leading-tight text-opacity-50 mt-2 sm:mt-0"
-      >
-        Join by scanning a QR with Bitkit, or by creating a new pubky.
-      </Typography.Body>
-      <div className="w-full flex-col inline-flex sm:grid sm:grid-cols-2 lg:grid-cols-8 gap-6 mt-6">
-        <Card.Join />
-        <Card.NewAccount />
-        <Card.SigninBack />
-      </div>
-      {/**<Content.MainBg alt="Onboard Pubky" imgSrc="/images/webp/bg-image1.webp" />*/}
-    </Onboarding.Layout>
-  );
+  return <SignIn.Content />;
 }
