+++ apps/web/app/profile/components/Handle/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { twMerge } from 'tailwind-merge';
 import { Typography } from '@social/ui-shared';
 import { useEffect, useState } from 'react';
