+++ apps/web/app/profile/components/_FilterTabs.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState, useEffect } from 'react';
 import { Icon, Typography } from '@social/ui-shared';
 import { Profile } from './';
