+++ apps/web/app/profile/components/_NotificationsProfile.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Notifications } from '@/app/notifications/components';
 import { Skeleton } from '@/components';
 import { useFilterContext } from '@/contexts';
