+++ apps/web/app/profile/components/index.tsx
@@ -4,6 +4,7 @@ import FilterTabs from './_FilterTabs';
 import FilterTabsMobile from './_FilterTabsMobile';
 import NotificationsProfile from './_NotificationsProfile';
 import Bookmarks from './_Bookmarks';
+import Content from './_Content';
 import Posts from './_Posts';
 import Replies from './_Replies';
 import Sidebar from './Sidebar';
@@ -15,6 +16,7 @@ export const Profile = {
   FilterTabsMobile,
   NotificationsProfile,
   Bookmarks,
+  Content,
   Posts,
   Replies,
   Sidebar,
